# Improve DictionaryBlock#copyPositions performance