--- a/core/trino-main/src/test/java/io/trino/operator/project/BenchmarkDictionaryBlockGetSizeInBytes.java
+++ b/core/trino-main/src/test/java/io/trino/operator/project/BenchmarkDictionaryBlockGetSizeInBytes.java
@@ -115,7 +115,7 @@ public class BenchmarkDictionaryBlockGetSizeInBytes
         {
             ImmutableList.Builder<String> list = ImmutableList.builder();
             for (int i = 0; i < count; i++) {
-                list.add(prefix + Integer.toString(i));
+                list.add(prefix + i);
             }
             return list.build();
         }