--- a/core/trino-spi/src/main/java/io/trino/spi/block/DictionaryBlock.java
+++ b/core/trino-spi/src/main/java/io/trino/spi/block/DictionaryBlock.java
@@ -18,9 +18,7 @@ import io.airlift.slice.Slices;
 import org.openjdk.jol.info.ClassLayout;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.function.BiConsumer;
 
 import static io.airlift.slice.SizeOf.sizeOf;
@@ -325,17 +323,18 @@ public class DictionaryBlock
         checkArrayRange(positions, offset, length);
 
         IntArrayList positionsToCopy = new IntArrayList();
-        Map<Integer, Integer> oldIndexToNewIndex = new HashMap<>();
+        Int2IntOpenHashMap oldIndexToNewIndex = new Int2IntOpenHashMap(min(length, dictionary.getPositionCount()));
         int[] newIds = new int[length];
 
         for (int i = 0; i < length; i++) {
             int position = positions[offset + i];
             int oldIndex = getId(position);
-            if (!oldIndexToNewIndex.containsKey(oldIndex)) {
-                oldIndexToNewIndex.put(oldIndex, positionsToCopy.size());
+            int newId = oldIndexToNewIndex.putIfAbsent(oldIndex, positionsToCopy.size());
+            if (newId == -1) {
+                newId = positionsToCopy.size();
                 positionsToCopy.add(oldIndex);
             }
-            newIds[i] = oldIndexToNewIndex.get(oldIndex);
+            newIds[i] = newId;
         }
         return new DictionaryBlock(dictionary.copyPositions(positionsToCopy.elements(), 0, positionsToCopy.size()), newIds);
     }