```
AFTER

Benchmark                               (selectedPositions)  Mode  Cnt      Score     Error  Units
BenchmarkDictionaryBlock.copyPositions                  100  avgt   45     38,054 ±   1,955  us/op
BenchmarkDictionaryBlock.copyPositions                 1000  avgt   45    271,195 ±   6,334  us/op
BenchmarkDictionaryBlock.copyPositions                10000  avgt   45   2248,107 ±  61,587  us/op
BenchmarkDictionaryBlock.copyPositions               100000  avgt   45  15093,510 ± 215,953  us/op

BEFORE

Benchmark                               (selectedPositions)  Mode  Cnt      Score     Error  Units
BenchmarkDictionaryBlock.copyPositions                  100  avgt   45     41,533 ±   2,254  us/op
BenchmarkDictionaryBlock.copyPositions                 1000  avgt   45    318,446 ±   8,066  us/op
BenchmarkDictionaryBlock.copyPositions                10000  avgt   45   2867,513 ±  50,908  us/op
BenchmarkDictionaryBlock.copyPositions               100000  avgt   45  20631,614 ± 546,668  us/op
```